+++ changelog.md
@@ -1,5 +1,6 @@
 # AvrixLoader v1.5.0
 
+- Added the ability to block the loading of lua files (useful for replacing Lua with Java)
 - Fixed visibility area of fields, classes and methods in Widgets
 - Fixed bugs when scrolling widgets
 - Fixed a bug with absolute positioning of widgets